ScrollBot/requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai


Flask
python-dotenv
google-generativeai
langchain
PyPDF2
faiss-cpu
